feat(6483): make a singular resultsView, whereas we have a separate results context (which we can then have multiple)
