+++ src/dataExplorer/components/Results.tsx
@@ -11,6 +11,7 @@ import {
 
 import {RemoteDataState, SimpleTableViewProperties} from 'src/types'
 import {ResultsContext} from 'src/dataExplorer/context/results'
+import {ResultsViewContext} from 'src/dataExplorer/context/resultsView'
 import {SidebarContext} from 'src/dataExplorer/context/sidebar'
 import {PersistanceContext} from 'src/dataExplorer/context/persistance'
 import {SearchWidget} from 'src/shared/components/search_widget/SearchWidget'
@@ -72,7 +73,8 @@ const EmptyResults: FC = () => {
 }
 
 const WrappedOptions: FC = () => {
-  const {result, view, setView} = useContext(ResultsContext)
+  const {result} = useContext(ResultsContext)
+  const {view, setView} = useContext(ResultsViewContext)
 
   return (
     <ViewOptions
@@ -94,7 +96,8 @@ const WrappedOptions: FC = () => {
 const Results: FC = () => {
   const [search, setSearch] = useState('')
   const {range} = useContext(PersistanceContext)
-  const {result, status, view, setView} = useContext(ResultsContext)
+  const {result, status} = useContext(ResultsContext)
+  const {view, setView} = useContext(ResultsViewContext)
   const {launch} = useContext(SidebarContext)
   const res = useMemo(() => {
     if (view.state === 'graph' || !search.trim() || !result?.parsed) {
